exp: refactor base.py into refs.py and exceptions.py
