+++ dvc/repo/experiments/executor/local.py
@@ -9,7 +9,7 @@ from funcy import cached_property
 from dvc.scm import SCM
 from dvc.utils.fs import remove
 
-from ..base import (
+from ..refs import (
     EXEC_APPLY,
     EXEC_BASELINE,
     EXEC_BRANCH,
@@ -25,7 +25,8 @@ if TYPE_CHECKING:
 
     from dvc.repo import Repo
 
-    from ..base import ExpRefInfo, ExpStashEntry
+    from ..refs import ExpRefInfo
+    from ..stash import ExpStashEntry
 
 logger = logging.getLogger(__name__)
