+++ dvc/repo/experiments/refs.py
@@ -1,0 +1,63 @@
+from typing import Optional
+
+from .exceptions import InvalidExpRefError
+
+# Experiment refs are stored according baseline git SHA:
+#   refs/exps/01/234abcd.../<exp_name>
+EXPS_NAMESPACE = "refs/exps"
+EXPS_STASH = f"{EXPS_NAMESPACE}/stash"
+EXEC_NAMESPACE = f"{EXPS_NAMESPACE}/exec"
+EXEC_APPLY = f"{EXEC_NAMESPACE}/EXEC_APPLY"
+EXEC_CHECKPOINT = f"{EXEC_NAMESPACE}/EXEC_CHECKPOINT"
+EXEC_BRANCH = f"{EXEC_NAMESPACE}/EXEC_BRANCH"
+EXEC_BASELINE = f"{EXEC_NAMESPACE}/EXEC_BASELINE"
+EXEC_HEAD = f"{EXEC_NAMESPACE}/EXEC_HEAD"
+EXEC_MERGE = f"{EXEC_NAMESPACE}/EXEC_MERGE"
+
+
+class ExpRefInfo:
+
+    namespace = EXPS_NAMESPACE
+
+    def __init__(
+        self, baseline_sha: Optional[str] = None, name: Optional[str] = None
+    ):
+        self.baseline_sha = baseline_sha
+        self.name: str = name if name else ""
+
+    def __str__(self):
+        return "/".join(self.parts)
+
+    def __repr__(self):
+        baseline = f"'{self.baseline_sha}'" if self.baseline_sha else "None"
+        name = f"'{self.name}'" if self.name else "None"
+        return f"ExpRefInfo(baseline_sha={baseline}, name={name})"
+
+    @property
+    def parts(self):
+        return (
+            (self.namespace,)
+            + (
+                (self.baseline_sha[:2], self.baseline_sha[2:])
+                if self.baseline_sha
+                else ()
+            )
+            + ((self.name,) if self.name else ())
+        )
+
+    @classmethod
+    def from_ref(cls, ref: str):
+        try:
+            parts = ref.split("/")
+            if (
+                len(parts) < 2
+                or len(parts) == 3
+                or len(parts) > 5
+                or "/".join(parts[:2]) != EXPS_NAMESPACE
+            ):
+                raise InvalidExpRefError(ref)
+        except ValueError:
+            raise InvalidExpRefError(ref)
+        baseline_sha = parts[2] + parts[3] if len(parts) >= 4 else None
+        name = parts[4] if len(parts) == 5 else None
+        return cls(baseline_sha, name)
