+++ tests/func/experiments/executor/test_ssh.py
@@ -7,9 +7,9 @@ import pytest
 from dvc_ssh.tests.cloud import TEST_SSH_KEY_PATH, TEST_SSH_USER
 
 from dvc.fs.ssh import SSHFileSystem
-from dvc.repo.experiments.base import EXEC_HEAD, EXEC_MERGE
 from dvc.repo.experiments.executor.base import ExecutorInfo, ExecutorResult
 from dvc.repo.experiments.executor.ssh import SSHExecutor
+from dvc.repo.experiments.refs import EXEC_HEAD, EXEC_MERGE
 from tests.func.machine.conftest import *  # noqa, pylint: disable=wildcard-import
